cyber: fix cyber and drivers to build on gcc-8

+++ modules/drivers/radar/conti_radar/conti_radar_message_manager.cc
@@ -140,7 +140,7 @@ void ContiRadarMessageManager::Parse(const uint32_t message_id,
     // if period 1.5 large than base period, inc error_count
     const double period_multiplier = 1.5;
     if (it->second.real_period >
-        (static_cast<double>(it->second.period) * period_multiplier)) {
+        static_cast<int64_t>((static_cast<double>(it->second.period) * period_multiplier))) {
       it->second.error_count += 1;
     } else {
       it->second.error_count = 0;
